Refactored commands to live in their own component
