+++ src/App.css
@@ -3,6 +3,7 @@
 
 .App {
   height: 100%;
+  font-family: Avenir;
 }
 
 .App-layout {
@@ -17,72 +18,4 @@
   min-width: 357px;
   max-width: 357px;
   height: 100%;
-}
-
-.custom-modal {
-  width: 605px;
-  position: absolute;
-  top: 80px;
-  left: 50%;
-  right: auto;
-  bottom: auto;
-  border: 0px none;
-  background-color: #EEEEEE;
-  /* background: rgba(255, 255, 255, 0.65); */
-  overflow: hidden;
-  border-radius: 4px;
-  outline: none;
-  padding: 10px;
-  box-shadow: 0px 1px 1px rgba(0, 0, 0, 0.1);
-  margin-right: -50%;
-  transform: translate(-50%, 0px);
-}
-
-.custom-input {
-  font-size: 14px;
-  border-radius: 4px;
-  border: 1px solid #EEEEEE;
-  box-shadow: 0px 1px 1px rgba(0, 0, 0, 0.1);
-  width: 585px;
-  padding: 6px;
-  outline: none;
-  color: #333333;
-  caret-color: #568af2;
-}
-
-.custom-inputFocused {
-  font-size: 14px;
-  border-radius: 4px;
-  border: 2px solid #568af2;
-  box-shadow: 0px 1px 1px rgba(0, 0, 0, 0.1);
-  width: 585px;
-  padding: 6px;
-  outline: none;
-  color: #333333;
-  caret-color: #568af2;
-}
-
-.custom-suggestion {
-  color: #333333;
-  font-weight: normal;
-  background: #FFFFFF;
-  padding: 14px 12px;
-}
-
-.custom-suggestionHighlighted {
-  font-weight: bold;
-  background-color: #FAFAFA;
-  padding: 14px 12px;
-}
-
-.custom-suggestionsContainer {
-  border: 1px solid #EEEEEE;
-  width: 585px;
-  border-radius: 4px;
-  box-shadow: 0px 1px 1px rgba(0, 0, 0, 0.1);
-  background: #FFFFFF;
-}
-
-.custom-suggestionsContainerOpen {
-  margin-top: 10px;
 }
