+++ src/App.tsx
@@ -8,6 +8,7 @@ import moment from 'moment';
 import CommandPalette from 'react-command-palette';
 import { IDailySchedule } from './components/DailySchedule';
 import Asana from './components/Asana';
+import Commands from './components/Commands';
 
 const { Sider, Content } = Layout
 
@@ -61,39 +62,9 @@ const App = () => {
     project: 'Backlog'
   }
 
-  const commands = [{
-    name: 'Create Asana Task',
-    command: () => console.log('Create Asana Task')
-  }, {
-    name: 'Create Task',
-    command: () => console.log('Create Task')
-  }]
-
-  const theme = {
-    container: 'atom-container',
-    containerOpen: 'atom-containerOpen',
-    content: 'atom-content',
-    header: 'atom-header',
-    input: 'custom-input',
-    inputFocused: 'custom-inputFocused',
-    inputOpen: 'atom-inputOpen',
-    modal: 'custom-modal',
-    overlay: 'atom-overlay',
-    spinner: 'atom-spinner',
-    suggestion: 'custom-suggestion',
-    suggestionFirst: 'atom-suggestionFirst',
-    suggestionHighlighted: 'custom-suggestionHighlighted',
-    suggestionsContainer: 'custom-suggestionsContainer',
-    suggestionsContainerOpen: 'custom-suggestionsContainerOpen',
-    suggestionsList: 'atom-suggestionsList',
-    trigger: 'atom-trigger'
-  }
-
   return (
     <div className="App">
-      <div style={{ position: 'absolute', bottom: '25px', left: '25px' }}>
-        <CommandPalette theme={theme} trigger={<Button className="App-command-palette-button">Command Palette</Button>} commands={commands} />
-      </div>
+      <Commands />
       <Layout className="App-layout">
         <Content className="App-content">
           <Schedule dailySchedules={schedules}/>
